jarvis-web/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
passlib
python-dotenv
edge-tts
aiofiles
requests
python-multipart